cleanup: move classes to their own directory
