+++ lib/classes/AsyncAF.js
@@ -1,5 +1,5 @@
 import {AsyncAfWrapperProto as AsyncAfProto} from './AsyncAfWrapper';
-import {staticMethods, prototypeMethods} from '../packageList';
+import {staticMethods, prototypeMethods} from '../../packageList';
 
 class AsyncAF extends AsyncAfProto {}
