+++ lib/classes/AsyncAfWrapper.js
@@ -1,4 +1,4 @@
-import use from './methods/other/use';
+import use from '../methods/other/use';
 
 class AsyncAfWrapperProto {
   constructor(data) {
